+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/contracts/ArgumentNullException.java
@@ -1,7 +1,7 @@
 package nl.ou.refactoring.advice.contracts;
 
 import java.text.MessageFormat;
-import java.util.ResourceBundle;
+import nl.ou.refactoring.advice.resources.ResourceProvider;
 
 /**
  * An exception that is thrown if an argument is null.
@@ -40,7 +40,10 @@ public final class ArgumentNullException extends NullPointerException {
 	 */
 	@Override
 	public String getLocalizedMessage() {
-		final var messageFormat = ResourceBundle.getBundle("ExceptionMessages").getString("argumentNull");
+		final var messageFormat =
+				ResourceProvider
+					.ExceptionMessages
+					.getMessageTemplate(this.getClass());
 		return MessageFormat.format(messageFormat, this.parameterName);
 	}
 }