+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/io/plantuml/GraphPlantUmlWriter.java
@@ -7,6 +7,9 @@ import nl.ou.refactoring.advice.contracts.ArgumentNullException;
 import nl.ou.refactoring.advice.io.GraphStringWriter;
 
 public abstract class GraphPlantUmlWriter extends GraphStringWriter {
+	private static final String START_UML = "@startuml";
+	private static final String END_UML = "@enduml";
+	
 	/**
 	 * Initialises a new instance of {@link GraphPlantUmlWriter}.
 	 * @param stringWriter Writes text output. Cannot be null.
@@ -19,4 +22,12 @@ public abstract class GraphPlantUmlWriter extends GraphStringWriter {
 	@Override
 	public abstract void write(Graph graph)
 			throws ArgumentNullException, GraphPathSegmentInvalidException;
+	
+	protected final void writeStartUml(String diagramName) {
+		this.printLine(String.format("%s %s", START_UML, diagramName));
+	}
+	
+	protected final void writeEndUml() {
+		this.printLine(END_UML);
+	}
 }