+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/workflow/risks/GraphNodeRiskBrokenSubTyping.java
@@ -34,9 +34,4 @@ public final class GraphNodeRiskBrokenSubTyping extends GraphNodeRisk {
 				(source, destination) -> new GraphEdgeAffects(source, destination),
 				GraphEdgeAffects.class);
 	}
-
-	@Override
-	public String getCaption() {
-		return "Broken Sub-typing";
-	}
 }