+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/workflow/risks/GraphNodeRiskCorrespondingSubclassSpecification.java
@@ -34,9 +34,4 @@ public final class GraphNodeRiskCorrespondingSubclassSpecification extends Graph
 				(source, destination) -> new GraphEdgeAffects(source, destination),
 				GraphEdgeAffects.class);
 	}
-
-	@Override
-	public String getCaption() {
-		return "Corresponding Subclass Specification";
-	}
 }