+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/workflow/risks/GraphNodeRiskForcedOverride.java
@@ -6,24 +6,24 @@ import nl.ou.refactoring.advice.edges.workflow.GraphEdgeAffects;
 import nl.ou.refactoring.advice.nodes.code.GraphNodeOperation;
 
 /**
- * Represents a "Missing Super Implementation" risk in a Refactoring Advice Graph.
- * This risk may arise if an overriding method is removed and none of the subclasses contain an alternative implementation.
+ * Represents a "Forced Override" risk in a Refactoring Advice Graph.
+ * This risk may arise if a method is defined in a superclass
+ * and adding a method with the same definition will introduce an override.
  */
-public final class GraphNodeRiskMissingSuperImplementation extends GraphNodeRisk {
+public final class GraphNodeRiskForcedOverride extends GraphNodeRisk {
 	/**
-	 * Initialises a new instance of {@link GraphNodeRiskMissingSuperImplementation}.
+	 * Initialises a new instance of {@link GraphNodeRiskForcedOverride}.
 	 * @param graph {@link Graph} The graph that contains the risk.
 	 * @throws ArgumentNullException Thrown if graph is null.
 	 */
-	public GraphNodeRiskMissingSuperImplementation(Graph graph)
-			throws ArgumentNullException {
+	public GraphNodeRiskForcedOverride(Graph graph) throws ArgumentNullException {
 		super(graph);
 	}
-
+	
 	/**
-	 * Indicates that the "Missing Super Implementation" affects an Operation.
+	 * Indicates that the "Forced Override" risk affects an Operation.
 	 * @param operationNode The affected Operation.
-	 * @return The edge that connects the "Missing Super Implementation" risk and the affected Operation.
+	 * @return The edge that connects the "Forced Override" risk and the affected Operation.
 	 * @throws ArgumentNullException Thrown if operationNode is null.
 	 */
 	public GraphEdgeAffects affects(GraphNodeOperation operationNode)