+++ nl.ou.refactoring.advice/src/main/java/nl/ou/refactoring/advice/nodes/workflow/risks/GraphNodeRiskMissingAbstractImplementation.java
@@ -34,9 +34,4 @@ public final class GraphNodeRiskMissingAbstractImplementation extends GraphNodeR
 				(source, destination) -> new GraphEdgeAffects(source, destination),
 				GraphEdgeAffects.class);
 	}
-
-	@Override
-	public String getCaption() {
-		return "Missing Abstract Implementation";
-	}
 }